+++ components/Fetch/FetchNyhende.tsx
@@ -1,0 +1,26 @@
+import { useEffect, useState } from "react";
+import { client as sanityClient } from "../../sanity/lib/client";
+
+export function useFetchPosts() {
+  const [posts, setPosts] = useState([]);
+
+  useEffect(() => {
+    async function fetchPosts() {
+      const query = `*[_type == "post" && "Nyhende" in categories[]->title]{
+        title,
+        slug,
+        author->{
+          name
+        },
+        mainImage,
+        publishedAt,
+        body
+      }`;
+      const result = await sanityClient.fetch(query);
+      setPosts(result);
+    }
+    fetchPosts();
+  }, []);
+
+  return posts;
+}
